Solar Splash 2022
Two files, with one loaded (DAQ_2022) to a microcontroller on the boat and one on land (Receiver_2022) connected to a computer. Hardware design involves 2 resistors connected in series to the battery set to create a voltage drop to calculate charge and discharge power and battery voltage and %. Includes GPS functions that were commented out because they did not return any non-zero values. Seems like the module never locked onto any satellites. Arduino connects to an LCD to display values, which are also transmitted via nRF24L01 SPI module (unsuccessfully :(). The receiver currently just outputs whatever it gets to serial monitor.